Address Flink connector review comments
