+++ flink-connector/v1.19/flink/src/main/java/org/apache/gravitino/flink/connector/iceberg/GravitinoIcebergCatalogFlink119.java
@@ -24,7 +24,7 @@ import org.apache.flink.table.catalog.AbstractCatalog;
 import org.apache.gravitino.flink.connector.PartitionConverter;
 import org.apache.gravitino.flink.connector.SchemaAndTablePropertiesConverter;
 import org.apache.gravitino.flink.connector.utils.CatalogCompat;
-import org.apache.gravitino.flink.connector.utils.CatalogCompatFlink119;
+import org.apache.gravitino.flink.connector.utils.DefaultCatalogCompat;
 import org.apache.iceberg.flink.FlinkCatalogFactory;
 
 /** {@link GravitinoIcebergCatalog} implementation for Flink 1.19. */
@@ -48,7 +48,7 @@ public class GravitinoIcebergCatalogFlink119 extends GravitinoIcebergCatalog {
 
   @Override
   protected CatalogCompat catalogCompat() {
-    return CatalogCompatFlink119.INSTANCE;
+    return DefaultCatalogCompat.INSTANCE;
   }
 
   private static AbstractCatalog createIcebergCatalog(
