+++ flink-connector/v1.19/flink/src/main/java/org/apache/gravitino/flink/connector/jdbc/GravitinoJdbcCatalogFlink119.java
@@ -28,7 +28,7 @@ import org.apache.flink.table.factories.Factory;
 import org.apache.gravitino.flink.connector.PartitionConverter;
 import org.apache.gravitino.flink.connector.SchemaAndTablePropertiesConverter;
 import org.apache.gravitino.flink.connector.utils.CatalogCompat;
-import org.apache.gravitino.flink.connector.utils.CatalogCompatFlink119;
+import org.apache.gravitino.flink.connector.utils.DefaultCatalogCompat;
 
 /**
  * {@link GravitinoJdbcCatalog} implementation for Flink 1.19. Uses the relocated {@code
@@ -57,6 +57,6 @@ public class GravitinoJdbcCatalogFlink119 extends GravitinoJdbcCatalog {
 
   @Override
   protected CatalogCompat catalogCompat() {
-    return CatalogCompatFlink119.INSTANCE;
+    return DefaultCatalogCompat.INSTANCE;
   }
 }
