+++ flink-connector/v1.19/flink/src/main/java/org/apache/gravitino/flink/connector/paimon/GravitinoPaimonCatalogFlink119.java
@@ -23,7 +23,7 @@ import org.apache.flink.table.factories.CatalogFactory;
 import org.apache.gravitino.flink.connector.PartitionConverter;
 import org.apache.gravitino.flink.connector.SchemaAndTablePropertiesConverter;
 import org.apache.gravitino.flink.connector.utils.CatalogCompat;
-import org.apache.gravitino.flink.connector.utils.CatalogCompatFlink119;
+import org.apache.gravitino.flink.connector.utils.DefaultCatalogCompat;
 
 /** {@link GravitinoPaimonCatalog} implementation for Flink 1.19. */
 public class GravitinoPaimonCatalogFlink119 extends GravitinoPaimonCatalog {
@@ -38,6 +38,6 @@ public class GravitinoPaimonCatalogFlink119 extends GravitinoPaimonCatalog {
 
   @Override
   protected CatalogCompat catalogCompat() {
-    return CatalogCompatFlink119.INSTANCE;
+    return DefaultCatalogCompat.INSTANCE;
   }
 }
